fix: move detection code to utils folder
